@@ -1,5 +1,5 @@
 module github.com/hattya/git-utime
 
-go 1.18
+go 1.19
 
 require golang.org/x/sys v0.5.0
