@@ -1,13 +1,12 @@
 //
 // git-utime :: utime_unix.go
 //
-//   Copyright (c) 2021 Akinori Hattori <hattya@gmail.com>
+//   Copyright (c) 2021-2023 Akinori Hattori <hattya@gmail.com>
 //
 //   SPDX-License-Identifier: MIT
 //
 
-//go:build !plan9 && !windows
-// +build !plan9,!windows
+//go:build unix
 
 package main
 
